Install Keycloak (#129)
* Add Docker Hub token.

* Pass "cluster" as an arg when generating YAML.

Cluster data is needed in many files so having it there is handy.

* Set Docker Hub username.

While GitHub's container registry ignores the username when using a
token, Docker does require the username to match the owner of the token.

* Revamp dependencies.

Allow more than just app & config to be declared: e.g. a databes should be possible, where the app depends on the database.

* Fix IPv6 node networking, again.

* Install Keycloak with CloudNative PG managed database.